Crate debris_std

The standard library implementation for the debris language

Currently experimental. There are not yet specific plans how it will look like.

However, I plan to add at least a wrapper for every minecraft command.

Macros

match_parameters

register_primitives

Functions

bool_to_int

dyn_int

execute

execute_format_string

execute_string

Executes a string as a command and returns the result

function_for

int_to_int

load

Loads the standard library module

load_all

print

print_format_string

print_int

print_int_static

print_string

register_primitives

Registers all primitive types

register_ticking_function

static_bool_to_int

static_int_to_int