Crate debris_std

The standard library implementation for the debris language

Currently experimental. There are not yet specific plans how it will look like.

However, I plan to add at least a wrapper for every minecraft command.

Macros

match_parameters ğŸ”’

register_primitives ğŸ”’

Functions

bool_to_int ğŸ”’

dbg ğŸ”’

dyn_int ğŸ”’

execute ğŸ”’

execute_format_string ğŸ”’

execute_string ğŸ”’

Executes a string as a command and returns the result

export ğŸ”’

Exports a function to a specific path

export_impl ğŸ”’

function_for ğŸ”’

int_to_int ğŸ”’

load

Loads the standard library module

load_all

on_tick ğŸ”’

Registers a function to run on load

print ğŸ”’

register_primitives ğŸ”’

Registers all primitive types

static_bool_to_int ğŸ”’

static_int_to_int ğŸ”’